# docs/04-技术架构/当前版本/V1-openapi.yaml
openapi: 3.0.3
info:
  title: DGOS V1 Logical API
  version: "1.0.0"
  description: |
    DGOS V1 逻辑契约的 HTTPS JSON + SSE 投影。桌面 IPC 必须保持相同的动作、字段、权限和事件语义。
    该文件同时登记已实现的本地 contract projection 与尚未具备真实桌面/Web E2E 证据的 draft operation。
security:
  - dgosSession: []
  - dgosApiKey: []
servers:
  - url: https://localhost
    description: Docker Compose Web projection (planned)
...
paths:
...
  /api/v1/apps:
    get:
      tags: [apps]
      operationId: listApps
      x-tag-feature: V1-FR-002
      x-tag-implementation-status: implemented-local-contract
      responses:
        '200':
          description: Public app catalog, only verified official or administrator-approved packages
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/AppPackageRecord' } } } }
    post:
      tags: [apps]
      operationId: submitAppManifest
      x-tag-feature: V1-FR-002
      x-tag-implementation-status: implemented-local-contract
      description: Verify real signed package bytes; caller cannot provide source, catalogState or effective trust. Scope app.catalog.manage; authentication and CSRF apply.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppPackageEnvelope' }
      responses:
        '201':
          description: Verified package staged for review or trusted official catalog
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppPackageRecord' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { description: Configured file-count or decoded/encoded request-size limit exceeded }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/apps/{appId}:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    get:
      operationId: getApp
      x-tag-feature: V1-FR-002
      parameters:
        - { in: query, name: version, schema: { type: string } }
        - { in: query, name: build, schema: { type: integer, minimum: 0 } }
      responses:
        '200': { $ref: '#/components/responses/AppPackage' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/apps/{appId}/test-install:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    post:
      operationId: testInstallApp
      x-tag-feature: V1-FR-002
      description: Separate app.package.test_install permission; does not approve catalog entry.
      requestBody: { $ref: '#/components/requestBodies/AppReleaseMutation' }
      responses:
        '200': { $ref: '#/components/responses/AppInstall' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/apps/{appId}/install:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    post:
      tags: [apps]
      operationId: installApp
      x-tag-feature: V1-FR-002
      x-tag-implementation-status: implemented-local-contract
      requestBody: { $ref: '#/components/requestBodies/AppReleaseMutation' }
      responses:
        '200': { $ref: '#/components/responses/AppInstall' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/apps/{appId}/launch:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    post:
      tags: [apps]
      operationId: launchApp
      x-tag-feature: V1-FR-001
      x-tag-implementation-status: implemented-local-contract
      requestBody: { $ref: '#/components/requestBodies/AppMutation' }
      responses:
        '200':
          description: Authenticated subject/session-bound isolated launch instance; declarations are not runtime grants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppLaunchReceipt' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/apps/{appId}/deployment:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    get:
      operationId: getAppDeployment
      x-tag-feature: V1-FR-002
      x-tag-implementation-status: planned
      description: app.catalog.read plus authenticated subject ownership; no subjectId override. Sole deployment lookup; /apps/installations is not a public alternative. Distinct from catalog reviewVersion. Return retained uninstalled record with activeRelease=null if present, otherwise 404 without enumerating other subjects. D uses versionNumber as update/uninstall baseVersion.
      responses:
        '200':
          description: Current subject deployment and optimistic concurrency revision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppDeployment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/apps/{appId}/bridge:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    post:
      operationId: invokeAppCapability
      security: [{ dgosSession: [] }]
      x-tag-feature: V1-FR-002
      x-tag-implementation-status: planned
      description: Authenticated host Session and CSRF only; recheck session/subject/instance/app/digest, installed declarations and current broker before registered handler. Capability inputs follow V1-app-bridge.schema.json call definitions. SDK names are not HTTP scopes; use the fixed mapping in the SDK contract. No arbitrary URL/method/headers or caller identity forwarding.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppCapabilityRequest' }
      responses:
        '200':
          description: Selected capability-specific redacted public response
          content:
            application/json:
              schema: { type: object }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/apps/{appId}/update:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    post:
      tags: [apps]
      operationId: updateApp
      x-tag-feature: V1-FR-002
      x-tag-implementation-status: implemented-local-contract
      requestBody: { $ref: '#/components/requestBodies/AppReleaseMutation' }
      responses:
        '200': { $ref: '#/components/responses/AppInstall' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/apps/{appId}/uninstall:
    parameters:
      - { $ref: '#/components/parameters/AppId' }
    post:
      tags: [apps]
      operationId: uninstallApp
      x-tag-feature: V1-FR-002
      x-tag-implementation-status: implemented-local-contract
      description: Removes registration/code pointer only, retains user data. No deleteData override is accepted.
      requestBody: { $ref: '#/components/requestBodies/AppMutation' }
      responses:
        '200': { $ref: '#/components/responses/AppInstall' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  requestBodies:
    AppReleaseMutation:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AppReleaseMutation' }
    AppMutation:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AppMutation' }
  securitySchemes:
    appLaunchTicket:
      type: apiKey
      in: query
      name: launchTicket
      description: Opaque short-lived session/subject/app/package/instance-bound static ticket, redacted from logs; no bridge authority.
    dgosSession:
      type: apiKey
      in: cookie
      name: dgos_session
      description: Revocable DGOS administrator or subject session cookie.
    dgosApiKey:
      type: http
      scheme: bearer
      bearerFormat: DGOS-API-Key
      description: DGOS API Key; the secret is never returned after creation.
  responses:
...
    AppPackage:
      description: Verified package metadata with platform-computed trust and catalog state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AppPackageRecord' }
    AppInstall:
      description: Authorized install snapshot; no host filesystem path is exposed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AppInstallRecord' }
...
    Forbidden:
      description: Subject lacks the required capability or scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found within the caller's authorized scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Idempotency, version, reservation, or resource conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Request failed schema or domain validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  parameters:
    AppId: { in: path, name: appId, required: true, schema: { type: string } }
...
  schemas:
...
    AppDeployment:
      type: object
      additionalProperties: false
      required: [appId, state, versionNumber, dataRetained, activeRelease]
      properties:
        appId: { type: string }
        state: { type: string, enum: [installed, active, update_pending, health_check_failed, rolled_back, uninstall_requested, uninstalled] }
        versionNumber: { type: integer, minimum: 1, description: Deployment revision used as update/uninstall baseVersion; never catalog reviewVersion or package build }
        dataRetained: { type: boolean }
        activeRelease:
          type: object
          nullable: true
          required: [version, build, releaseChannel, digest, dataVersion]
          properties:
            version: { type: string }
            build: { type: integer, minimum: 0 }
            releaseChannel: { type: string, enum: [stable, beta, dev] }
            digest: { type: string }
            dataVersion: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
    AppLaunchReceipt:
      type: object
      required: [appId, instanceId, bridgeVersion, entrypoint, isolation, declaredCapabilities, expiresAt]
      properties:
        appId: { type: string }
        instanceId: { type: string, format: uuid }
        bridgeVersion: { type: integer, enum: [1] }
        entrypoint: { type: string, description: Controlled relative resource route; optional launchTicket is sensitive, do not log or persist }
        isolation: { type: string, enum: [opaque-origin-sandbox] }
        declaredCapabilities: { type: array, uniqueItems: true, items: { type: string }, description: Declared ceiling only, not runtime grants }
        expiresAt: { type: string, format: date-time }
        digest: { type: string }
    AppCapabilityRequest:
      type: object
      additionalProperties: false
      required: [instanceId, requestId, capability, input]
      properties:
        instanceId: { type: string, format: uuid }
        requestId: { type: string, format: uuid }
        capability: { type: string, enum: [dgos.model.list, dgos.model.resolve, dgos.aiTask.submit, dgos.aiTask.get, dgos.aiTask.events, dgos.aiTask.cancel, dgos.artifact.read, dgos.system.context.read, dgos.system.context.events] }
        input: { type: object, description: Strict selected capability schema; no caller owner/session/app or transport URL/handler }
      x-dgos-call-schema: 'V1-app-bridge.schema.json#/$defs/call'
      x-dgos-capability-mapping:
        dgos.model.list: { scopes: [provider.model.read], response: ModelSelectionList }
        dgos.model.resolve: { scopes: [provider.model.read], response: ModelResolution }
        dgos.aiTask.submit: { scopes: [ai_task.submit], operationId: submitAiTask, response: AITaskReceipt }
        dgos.aiTask.get: { scopes: [ai_task.read], operationId: getAiTask, response: AITaskSnapshot }
        dgos.aiTask.events: { scopes: [ai_task.read], operationId: streamAiTaskEvents, response: AppTaskEventBatch }
        dgos.aiTask.cancel: { scopes: [ai_task.cancel], operationId: cancelAiTask, response: AITaskSnapshot }
        dgos.artifact.read: { scopes: [artifact.read], operationId: readArtifact, response: ArtifactContent }
        dgos.system.context.read: { scopes: [dgos.system.context.read], response: AppSystemContext }
        dgos.system.context.events: { scopes: [dgos.system.context.events, dgos.system.context.read], response: AppContextEventBatch }
...
    AppPackageEnvelope:
      type: object
      additionalProperties: false
      required: [requestId, manifest, files, resourceDigests, keyId, signature]
      description: Ed25519 signature over UTF-8 canonical JSON of manifest and resourceDigests. Operator-configured limits default to 1000 files and 32 MiB decoded content. HTTP encoded-body limit must account for base64 overhead and reject excess before processing.
      properties:
        requestId: { type: string, format: uuid }
        manifest:
          type: object
          description: Must validate against V1-app-manifest.schema.json (JSON Schema 2020-12); this object does not grant trust or catalog approval.
          externalDocs: { url: 'V1-app-manifest.schema.json' }
        files: { type: object, additionalProperties: { type: string, format: byte }, description: Normalized package-relative POSIX path to strict base64 bytes; no traversal, absolute paths, symlinks or installer scripts }
        resourceDigests: { type: object, additionalProperties: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }, description: Exactly the files key set, each digest recomputed from decoded bytes }
        keyId: { type: string, minLength: 1, description: Selects an already provisioned non-revoked operator trust root; cannot supply a root }
        signature: { type: string, format: byte, description: Ed25519 signature, strict base64 of 64 bytes }
    AppPackageRecord:
      type: object
      required: [appId, version, build, releaseChannel, digest, source, catalogState, trustLevel, uninstallPolicy]
      properties:
        appId: { type: string }
        version: { type: string }
        build: { type: integer, minimum: 0 }
        releaseChannel: { type: string, enum: [stable, beta, dev] }
        digest: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
        source: { type: string, enum: [official, admin, developer], readOnly: true }
        catalogState: { type: string, enum: [official, pending_review, approved, rejected, withdrawn], readOnly: true }
        trustLevel: { type: string, enum: [standard, trusted, system], readOnly: true, description: Effective platform grant, never copied from requested manifest trust }
        uninstallPolicy: { type: string, enum: [user-removable, protected-preinstall], readOnly: true }
        manifest: { type: object, description: Original signed manifest; trustLevel inside is only the requested value }
        reviewVersion: { type: integer, minimum: 1 }
    AppReleaseMutation:
      type: object
      additionalProperties: false
      required: [requestId, version, build, releaseChannel]
      properties:
        requestId: { type: string, format: uuid }
        version: { type: string }
        build: { type: integer, minimum: 0 }
        releaseChannel: { type: string, enum: [stable, beta, dev] }
        baseVersion: { type: integer, minimum: 1, description: Required when mutating an existing versioned install/review record; checked again under the installation/review lock }
        reason: { type: string, description: Redacted administrator review rationale }
    AppMutation:
      type: object
      additionalProperties: false
      required: [requestId]
      properties:
        requestId: { type: string, format: uuid }
        baseVersion: { type: integer, minimum: 1 }
    AppInstallRecord:
      type: object
      required: [appId, state, dataRetained]
      properties:
        appId: { type: string }
        version: { type: string }
        build: { type: integer, minimum: 0 }
        releaseChannel: { type: string, enum: [stable, beta, dev] }
        state: { type: string, enum: [discovered, installed, active, update_pending, health_check_failed, rolled_back, uninstall_requested, uninstalled] }
        dataRetained: { type: boolean }
        digest: { type: string }
        previousDigest: { type: string, nullable: true }
        versionNumber: { type: integer, minimum: 1 }
    ErrorResponse:
      type: object
      required: [errorKey, message, requestId]
      properties:
        errorKey: { type: string, description: Stable machine-readable error key }
        message: { type: string, description: Safe human-readable message without secrets }
        requestId: { type: string }
        retryable: { type: boolean }
        details: { type: object, additionalProperties: true, description: Redacted diagnostic details }
...